Restructured crate a bit
