+++ src/prelude.rs
@@ -1,0 +1,2 @@
+pub use crate::core::{Limage, LimageError};
+pub use crate::patterns::{Grid, Pattern};
